requirements.txt
flask
python-gnupg
pytest
gunicorn